Remove hack from late_field_path

+++ tests/end_to_end/division-by-power-of-two/ido53-o2-out.c
@@ -1,3 +1,3 @@
 void test(s32 *arg0) {
-    *arg0 = (s32) ((s32) *arg0 / 2);
+    *arg0 = (s32) *arg0 / 2;
 }
