+++ tests/end_to_end/error/manual-out.c
@@ -18,7 +18,7 @@
     MIPS2C_TRAP_IF((u32) arg0 >= 6U);
     MIPS2C_ERROR(unknown instruction: badinstr $t0, $t0);
     temp_t1 = MIPS2C_ERROR(unknown instruction: badinstr2 $t1, $t1);
-    *NULL = (s32) (temp_t1 << temp_t1);
+    *NULL = temp_t1 << temp_t1;
     *NULL = (s32) (MIPS2C_ERROR(Read from unset register $v1) + 2);
     return MIPS2C_ERROR(unknown instruction: badinstr3 $v0, $t2);
 }
