+++ tests/end_to_end/if_postdec/irix-o2-out.c
@@ -4,7 +4,7 @@ extern s32 D_4100E0;
     s32 temp_v1;
 
     temp_v1 = D_4100E0;
-    D_4100E0 = (s32) (temp_v1 - 1);
+    D_4100E0 = temp_v1 - 1;
     if (temp_v1 <= 0) {
         return 4;
     }
